requirements.txt
Flask
flask-cors
requests
beautifulsoup4
lxml
python-dotenv
urllib3
gunicorn
supabase
